address initial code review
